tests/wpt/web-platform-tests/html/browsers/browsing-the-web/history-traversal/browsing_context_name_cross_origin.html

<!doctype html>
<title>Restoring window.name on cross-origin history traversal</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<pre id="step_log"></pre>

<script>
var t = async_test();
t.step(() => {
  var win = window.open("browsing_context_name_cross_origin-0.html");
  t.add_cleanup(() => win.close());
});
</script>
